Refactor search page into client component and localized routes
Extract the search logic into a shared client component and establish separate route pages for Norwegian and English locales. This enables the addition of metadata to prevent search engine indexing for search result pages, aligning with the Next.js migration strategy.
